@@ -5,14 +5,13 @@ import 'package:meta/meta.dart';
 class EpisodesActionFailed implements Exception {}
 
 class EpisodesRepository {
-  EpisodesRepository({@required this.apiRepository})
-      : assert(apiRepository != null);
+  EpisodesRepository({required this.apiRepository});
   final APIRepository apiRepository;
   Future<List<Episode>> getGamesEpisodes({int page = 1}) async {
     try {
       List<Episode> result = [];
-      List<dynamic> apiResponse = await apiRepository
-          .performGet('http://161.35.255.93:3000/episodes/games?page=$page');
+      List<dynamic> apiResponse = await apiRepository.performGet(
+          'https://68fc-177-66-113-212.ngrok.io/episodes/games?page=$page');
       result = apiResponse.map((e) => Episode.fromJson(e)).toList();
       return result;
     } catch (e) {
@@ -25,7 +24,7 @@ class EpisodesRepository {
     try {
       List<Episode> result = [];
       List<dynamic> apiResponse = await apiRepository.performGet(
-          'http://161.35.255.93:3000/episodes/non-games?page=$page');
+          'https://68fc-177-66-113-212.ngrok.io/episodes/non-games?page=$page');
       result = apiResponse.map((e) => Episode.fromJson(e)).toList();
       return result;
     } on Exception {
